requirements.txt
Flask
Flask-Cors
Flask-Migrate
flask-restx
Flask-SQLAlchemy
Flask-WTF
botocore
boto3
python-dotenv
json-logging
SQLAlchemy
pandas
APScheduler